save changes

+++ app/Http/Controllers/Admin/Content/CategoryController.php
@@ -9,10 +9,16 @@ use App\Http\Services\Image\ImageService;
 use App\Models\Content\PostCategory;
 use http\Env\Response;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 
 class CategoryController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('can:show-category')->only(['index']);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -20,6 +26,9 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        if(Auth::user()->cannot('show-category')){
+            return redirect()->back()->with('toast-error','دسترسی غیرمجاز');
+        }
         $postCategories = PostCategory::orderBy('created_at', 'desc')->simplePaginate(15);
         return view('admin.content.category.index', compact('postCategories'));
     }
@@ -31,6 +40,9 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        if(Auth::user()->cannot('create-category')){
+            return redirect()->back()->with('toast-error','دسترسی غیرمجاز');
+        }
         $categories = PostCategory::all();
         return view('admin.content.category.create', compact('categories'));
     }
@@ -43,6 +55,9 @@ class CategoryController extends Controller
      */
     public function store(PostCategoryRequest $request, ImageService $imageService)
     {
+        if(Auth::user()->cannot('create-category')){
+            return redirect()->back()->with('toast-error','دسترسی غیرمجاز');
+        }
         $inputs = $request->all();
         if ($request->hasFile('image')){
             /*$img = new ImageCacheService();
@@ -78,6 +93,9 @@ class CategoryController extends Controller
      */
     public function edit($id)
     {
+        if(Auth::user()->cannot('update-category')){
+            return redirect()->back()->with('toast-error','دسترسی غیرمجاز');
+        }
         $categories = PostCategory::all();
         $category = PostCategory::find($id);
         return view('admin.content.category.edit', compact('categories', 'category'));
@@ -92,6 +110,9 @@ class CategoryController extends Controller
      */
     public function update(PostCategoryRequest $request, $id, ImageService $imageService)
     {
+        if(Auth::user()->cannot('update-category')){
+            return redirect()->back()->with('toast-error','دسترسی غیرمجاز');
+        }
         $postCategory=PostCategory::find($id);
         $inputs = $request->all();
         if ($request->hasFile('image')){
@@ -125,6 +146,9 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        if(Auth::user()->cannot('delete-category')){
+            return redirect()->back()->with('toast-error','دسترسی غیرمجاز');
+        }
         $category = PostCategory::find($id);
         $category->delete();
         return redirect()->back()->with('swal-success', 'دسته بندی با موفقیت حذف شد');
